+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add support to specify project name either via `.dir-locals.el` or by providing a customized `projectile-project-name-function'.
 * Add a command to switch between open projects (`projectile-switch-open-project`).
 * Add a command to edit the .dir-locals.el file of the project (`projectile-edit-dir-locals`).
+* Add file local variable projectile-custom-root, which allows overriding the project root on a per-file basis. This allows navigating a different project from, say, an org file in a another git repository.
 
 ### Changes